KULRICE-3905: Modified KEW's email services and "mock" email services so that they function properly without the annotation-based routeHeader-actionItem link; added extra @Transient properties to DocumentRouteHeaderValue and ActionRequestValue for simulation engine purposes. KULRICE-3962: Fixed the RouteNodeInstance-related exception routing problem by ensuring that the newly-generated exception requests reference null node instances; removed annotation-based link between action requests and action items in order to fix lazy-loading problems related to the exception routing service, meaning a request's action items are now being obtained via service calls; added extra simulation-related @Transient property on action requests similar to those for KULRICE-3905. KULRICE-3862: Added findByActionRequestId() to the ActionListService interface; added a temporary fix for persisting ApplicationDocumentStatus instances on a DocumentType; added some temporary try-catch blocks for certain NoResultExceptions until getting feedback about whether to use an interceptor to handle them instead; changed the SUBSELECT fetch modes on eager KEW collections into SELECT fetch modes; Fixed certain KEW DAOs relying on Rice's Criteria API so that the JPA alias is now being used properly in their queries; Made some miscellaneous fixes to various entity annotations, DAOs, and related classes.

+++ impl/src/main/java/org/kuali/rice/kew/actiontaken/dao/impl/ActionTakenDAOJpaImpl.java
@@ -137,7 +137,7 @@ public class ActionTakenDAOJpaImpl implements ActionTakenDAO {
 	    crit.eq("routeHeaderId", routeHeaderId);
 	    crit.eq("principalId", workflowId);
 	    crit.eq("currentIndicator", Boolean.TRUE);
-	    int count = (Integer) new QueryByCriteria(entityManager, crit).toCountQuery().getSingleResult();
+	    long count = (Long) new QueryByCriteria(entityManager, crit).toCountQuery().getSingleResult();
         return count > 0;
     }
